compiles again but no error check yet

+++ ed25519/ed25519.c
@@ -29,7 +29,7 @@
 void ed25519_text2key(uint8_t* key,const char* text,int len) // len = key length in bytes
 {       int i;
         char x[3]="00";
-        assert(strlen(text)==2*len);
+        assert(strlen(text)==(size_t)2*len);
         for(i=0;i<len;i++){
                 x[0]=text[2*i+0];
                 x[1]=text[2*i+1];
